+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 9.2.0 IN PROGRESS
 
 * Use Save & close button label stripes-component translation key. Refs UICIRC-1066.
+* Remove bigtests from github actions. Refs UICIRC-1079.
 
 ## [9.1.0](https://github.com/folio-org/ui-circulation/tree/v9.1.0) (2024-03-22)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v9.0.5...v9.1.0)