+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * Use Save & close button label stripes-component translation key. Refs UICIRC-1066.
 * Remove bigtests from github actions. Refs UICIRC-1079.
-* Create new permission 'Settings (Circ): Can enable request print details'. Refs UICIRC-3153.
+* Create new permission 'Settings (Circ): Can enable request print details'. Refs UICIRC-1093.
 
 ## [9.1.0](https://github.com/folio-org/ui-circulation/tree/v9.1.0) (2024-03-22)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v9.0.5...v9.1.0)