+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Remove bigtests from github actions. Refs UICIRC-1079.
 * Create new permission 'Settings (Circ): Can enable request print details'. Refs UICIRC-1093.
 * Create a new circ settings page and configuration to enable request print history. Refs UICIRC-1087.
-* Add page title to Circulation settings - VIew Print Details Page. Refs UICIRC-1089.
+* Add page title to Circulation settings - View Print Details Page. Refs UICIRC-1089.
 
 ## [9.1.0](https://github.com/folio-org/ui-circulation/tree/v9.1.0) (2024-03-22)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v9.0.5...v9.1.0)